+++ oqs/__init__.py
@@ -1,7 +1,7 @@
-from oqs.oqswrap import _enabled_KEMs, _supported_KEMs, KeyEncapsulation
-from oqs.oqswrap import _enabled_sigs, _supported_sigs, Signature
-from oqs.oqswrap import MechanismNotEnabledError, MechanismNotSupportedError
+from oqs.wrapper import _enabled_KEMs, _supported_KEMs, KeyEncapsulation
+from oqs.wrapper import _enabled_sigs, _supported_sigs, Signature
+from oqs.wrapper import MechanismNotEnabledError, MechanismNotSupportedError
 
 # TODO: refactor this out
-from oqs.oqswrap import print_enabled_KEM_mechanisms, print_enabled_sig_mechanisms
-from oqs.oqswrap import get_enabled_KEM_mechanisms, get_enabled_sig_mechanisms
+from oqs.wrapper import print_enabled_KEM_mechanisms, print_enabled_sig_mechanisms
+from oqs.wrapper import get_enabled_KEM_mechanisms, get_enabled_sig_mechanisms
